requirements.txt
django
djangorestframework
django-cors-headers

vite: 
tailwindcss
headlessui
heroicons
rest_framework_simplejwt